v1.0.1

+++ database/migrations/2021_05_18_123249_create_products_table.php
@@ -10,7 +10,7 @@ class CreateProductsTable extends Migration{
       $table->id();
       $table->string('title');
       $table->text('description');
-      $table->integer('price')->unsigned();
+      $table->float('price')->unsigned();
       $table->string('currency')->default('GEL');
       $table->timestamps();
     });
